Update StudentSchedulePage.vue
Couldn't push from VS as my version of this page was older. but this new commit is the latest version of this page.
Worked on getting more data onto the calendar dates, changed colors based on the right conditions, and added cancel and register buttons in respectable sessions.
Now need to generate the upcoming session table then finalize the cancel and register buttons.